+++ packages/block-library/src/navigation/edit/menu-inspector-controls.js
@@ -132,7 +132,7 @@ const DefaultControls = ( props ) => {
 const MenuInspectorControls = ( props ) => {
 	// Show the OffCanvasEditor controls if we're in the Gutenberg plugin. Previously used isOffCanvasNavigationEditorEnabled.
 	return (
-		<InspectorControls __experimentalGroup="list">
+		<InspectorControls group="list">
 			<PanelBody
 				title={ process.env.IS_GUTENBERG_PLUGIN ? null : __( 'Menu' ) }
 			>