+++ packages/components/src/notice/test/index.tsx
@@ -14,14 +14,15 @@ import { speak } from '@wordpress/a11y';
 import Notice from '../index';
 
 jest.mock( '@wordpress/a11y', () => ( { speak: jest.fn() } ) );
+const mockedSpeak = jest.mocked( speak );
 
-function getNoticeWrapper( container ) {
+function getNoticeWrapper( container: HTMLElement ) {
 	return container.firstChild;
 }
 
 describe( 'Notice', () => {
 	beforeEach( () => {
-		speak.mockReset();
+		mockedSpeak.mockReset();
 	} );
 
 	it( 'should match snapshot', () => {
@@ -42,7 +43,9 @@ describe( 'Notice', () => {
 	} );
 
 	it( 'should not have is-dismissible class when isDismissible prop is false', () => {
-		const { container } = render( <Notice isDismissible={ false } /> );
+		const { container } = render(
+			<Notice isDismissible={ false }>I cannot be dismissed!</Notice>
+		);
 		const wrapper = getNoticeWrapper( container );
 
 		expect( wrapper ).toHaveClass( 'components-notice' );
@@ -50,7 +53,7 @@ describe( 'Notice', () => {
 	} );
 
 	it( 'should default to info status', () => {
-		const { container } = render( <Notice /> );
+		const { container } = render( <Notice>FYI</Notice> );
 
 		expect( getNoticeWrapper( container ) ).toHaveClass( 'is-info' );
 	} );